refactor: move test modules into dedicated files for diagram and knot-so-good
- src/diagram/test.rs  — extracted from inline mod test {} in diagram.rs
- src/diagram/tests.rs — extracted module-level #[test] fns from diagram.rs
- examples/knot-so-good/src/tests.rs — extracted from inline mod tests {} in main.rs
- Snapshot files relocated to src/diagram/snapshots/ to match new source path

Module hierarchy is unchanged: knotty::diagram::test, knotty::diagram::tests,
knot_so_good::tests resolve identically to before.
